requirement.txt
django==2.0.5
djangorestframework==3.8.2